Version 3.2
Beta version

Description:
Start refactoring the code, which will divide the code with functions among themselves,
which will greatly simplify the task of reading and editing the code.
The code in the file (the name of the file) will be called the same as the designation of what it does.

Added:
handlers/
	This folder will store various handlers for various commands that the bot responds to
handlers/__init__.py
	Directory initialization file
handlers/registration.py
	Added registration functions (do not work completely)

Modified:
main.py
	Registration functions moved to "handlers/registration.py"